Bump the OpenOCD and Picotool revision #s
The Arduino IDE will not even look inside the JSON for new files unless
the revision doesn't match the currently installed one.  So, in this case,
the IDE is not installing new OpenOCD or Picotool even though it's listed
in the package.json file.

Bump the revision ID such that the IDE darn well better install the newer
revisions.

Fixes #2355